Add GPU utilization metric
Also try to adjust cpu/mem metrics for multiple-node clusters.
But sum or average probably won't be the right approach here, we
probably would be looking for the max (or min) utilization.

Then again we'd have to start digging deeper into the metrics because it
might not be useful if one node happens to have lots of CPU, another has
lots of memory and a third happens to have the one free GPU...