Labels are restrictive with accepted characters
And we don't actually use the chart value for anything but presentation
because we pull the actual deployment info from the uuid anyway.
So instead of stripping out not-acceptable characters or running this
through urlsafe base64 encoding, just drop the value. If we really need
it we should store a compressed and encoded score (chart/version/values).

Also switch to using the example scores that are stored in the container.