Wait for helm chart to finish deployment
Currently this blocks the Flask thread for up to 5 minutes (default
timeout for helm), we could shorten the timeout but that could lead
to unnecessary deployment failures.

A longer term solution would be to run the helm deployment as a
background thread with something like Flask-Executor and monitor
progress/status of the deployment with a separate HTTP GET call.

This however will lead to additional changes in Tier1 which already
uses Flask-Executor to trigger parallel deployments but still blocks
until all deployments either complete or fail.