Add s3 upload helper
Run with `poetry run python -m sinfonia.s3_upload` for now.

Still have to handle secure access to S3. Clients know the UUID of the
backend and if they guess the bucket name they can access the recipes.

Right now we have to use an S3 (minio) instance that can only be
accessed by tier1/tier2, but that complicates updating the recipes.